util: expose getIp, add `sendGrcRequestPlain`
currently the only way to send a request with ip is by passing
in the request object from express. together with it the body/query
is read and put into a single object into `args` array.

this comes with multiple problems:

 - when here is the need to pass in extra values, which are not
   initially present in the request object, you have to monkey
   patch `req.body`

 - for services requiring more than one arg, there is no good way
   to add those.

thinking about https://github.com/bitfinexcom/bfx-lib-server-js/issues/16
it makes sense to expose the shared ip helper, so users do no need
to pass in the req object. this also adds a reqeust function that
supports sanitizing and is not bound to express request objects and
their content.

closes #16